Update documentation for message classes
