+++ Dockerfile
@@ -1,0 +1,2 @@
+FROM nginx
+ADD build/web /usr/share/nginx/html
