+++ lib/src/messages/control.dart
@@ -1,12 +1,19 @@
 part of client.src.messages.message;
 
+/// A control message represents messages communicated by the application and
+/// are emitted by changes within the state of the application. Control messages
+/// are unidireciton in that they only display to a single user and do not
+/// traverse the network.
 class ControlMessage extends Message {
-  String body;
-  final bool isControl = true;
   bool viewed = true;
 
-  ControlMessage(this.body);
+  ControlMessage(String messageText) {
+    body = messageText;
+  }
 
   @override
-  Map marshal() => {};
+  Map marshal() => {
+        // Becuase control messages are generated by the application they do not
+        // traverse the network so calls to marshal return no value.
+      };
 }
