+++ lib/src/messages/file.dart
@@ -1,24 +1,20 @@
 part of client.src.messages.message;
 
-class FileMessageFactory {
-  FileMessage createFileMessage(Uri uri, String contentType) {
-    if (contentType.startsWith("image/")) {
-      return new Image(uri, contentType);
-    } else if (contentType.startsWith("video/")) {
-      return new Video(uri, contentType);
-    } else {
-      return new FileMessage(uri, contentType);
-    }
-  }
-}
-
+/// A [FileMessage] is a specialized message that represents a single file item
+/// sent by a user.
 class FileMessage extends Message {
+  /// A human redable title for the file.
   String title;
+
+  /// The resource location for the file
   Uri uri;
+
+  /// The mime-type of the FileMessage.
   String contentType;
 
   FileMessage(this.uri, this.contentType);
 
+  /// Captures the message state as a map for serialization.
   @override
   Map marshal() => {
         'author': author.name,
@@ -29,10 +25,12 @@ class FileMessage extends Message {
       };
 }
 
+/// A specialized message type for Image based files.
 class Image extends FileMessage {
   Image(Uri uri, String contentType) : super(uri, contentType);
 }
 
+/// A specialized message type for Video based files.
 class Video extends FileMessage {
   Video(Uri uri, String contentType) : super(uri, contentType);
 }
