+++ lib/src/messages/link.dart
@@ -1,7 +1,11 @@
 part of client.src.messages.message;
 
+/// Represents a link shared by a message.
 class Link extends Message {
+  /// A human readable title for the link
   String title;
+
+  /// The resource identifier for the given link
   Uri ref;
 
   Map marshal() {
