+++ lib/src/web_application/presenters/presenters.dart
@@ -1,7 +1,5 @@
 library client.src.web_application.presenters.presenters;
 
-import 'package:markdown/markdown.dart';
-
 import 'package:client/src/messages/presenter.dart';
 import 'package:client/src/messages/message.dart';
 
@@ -9,6 +7,7 @@ part 'control_presenter.dart';
 part 'file_presenters.dart';
 part 'message_presenter.dart';
 
+/// A simplified factory for returning a presenter object for a given message.
 class SimplePresenterFactory extends PresenterFactory {
   Presenter getPresenter(Message m) {
     if (m is MarkdownMessage) {
@@ -20,6 +19,7 @@ class SimplePresenterFactory extends PresenterFactory {
     } else if (m is FileMessage) {
       return new FilePresenter(m);
     } else {
+      /// All non specialized messages are rendered as ControlMessages
       return new ControlPresenter(m);
     }
   }
