+++ web/main.dart
@@ -2,8 +2,9 @@ import 'package:angular/application_factory.dart';
 import 'package:client/web_application.dart';
 
 void main() {
+  // Create a new Angular Dart application
   applicationFactory()
-      .rootContextType(ImUniClient)
+      .rootContextType(UniImClientFacade)
       .addModule(new UniImModule())
       .run();
 }
